proper class for tests
