+++ boards/DIM/Test/Inc/Test_Dim.h
@@ -8,21 +8,3 @@
 
 #include <fff.h>
 #include <gtest/gtest.h>
-
-extern "C"
-{
-#include "App_SevenSegDisplay.h"
-#include "App_SevenSegDisplays.h"
-}
-
-class DimTest : public testing::Test
-{
-  protected:
-    struct SevenSegDisplay * left_seven_seg_display;
-    struct SevenSegDisplay * middle_seven_seg_display;
-    struct SevenSegDisplay * right_seven_seg_display;
-    struct SevenSegDisplays *seven_seg_displays;
-
-    void SetUp() override;
-    void TearDown() override;
-};
