+++ boards/DIM/Test/Src/Test_Dim.cpp
@@ -1,36 +1,3 @@
 #include "Test_Dim.h"
-#include "Test_SevenSegDisplays.h"
 
 DEFINE_FFF_GLOBALS;
-
-void DimTest::SetUp()
-{
-    left_seven_seg_display   = App_SevenSegDisplay_Create(set_left_hex_digit);
-    middle_seven_seg_display = App_SevenSegDisplay_Create(set_middle_hex_digit);
-    right_seven_seg_display  = App_SevenSegDisplay_Create(set_right_hex_digit);
-    seven_seg_displays       = App_SevenSegDisplays_Create(
-        left_seven_seg_display, middle_seven_seg_display,
-        right_seven_seg_display);
-
-    RESET_FAKE(set_right_hex_digit);
-    RESET_FAKE(set_middle_hex_digit);
-    RESET_FAKE(set_left_hex_digit);
-}
-
-void DimTest::TearDown()
-{
-    ASSERT_TRUE(left_seven_seg_display != NULL);
-    ASSERT_TRUE(middle_seven_seg_display != NULL);
-    ASSERT_TRUE(right_seven_seg_display != NULL);
-    ASSERT_TRUE(seven_seg_displays != NULL);
-
-    App_SevenSegDisplay_Destroy(left_seven_seg_display);
-    App_SevenSegDisplay_Destroy(middle_seven_seg_display);
-    App_SevenSegDisplay_Destroy(right_seven_seg_display);
-    App_SevenSegDisplays_Destroy(seven_seg_displays);
-
-    left_seven_seg_display   = NULL;
-    middle_seven_seg_display = NULL;
-    right_seven_seg_display  = NULL;
-    seven_seg_displays       = NULL;
-}
