+++ boards/DIM/Test/Src/Test_SevenSegDisplays.cpp
@@ -3,12 +3,39 @@
 DEFINE_FAKE_VOID_FUNC(set_right_hex_digit, struct SevenSegHexDigit);
 DEFINE_FAKE_VOID_FUNC(set_middle_hex_digit, struct SevenSegHexDigit);
 DEFINE_FAKE_VOID_FUNC(set_left_hex_digit, struct SevenSegHexDigit);
+DEFINE_FAKE_VOID_FUNC(set_value_callback);
 
-class SevenSegDisplaysTest : public DimTest
+void SevenSegDisplaysTest::SetUp()
 {
-    // Empty class because we just re-use the Setup() and Teardown() from the
-    // base class.
-};
+    left_seven_seg_display   = App_SevenSegDisplay_Create(set_left_hex_digit);
+    middle_seven_seg_display = App_SevenSegDisplay_Create(set_middle_hex_digit);
+    right_seven_seg_display  = App_SevenSegDisplay_Create(set_right_hex_digit);
+    seven_seg_displays       = App_SevenSegDisplays_Create(
+        left_seven_seg_display, middle_seven_seg_display,
+        right_seven_seg_display, set_value_callback);
+
+    RESET_FAKE(set_right_hex_digit);
+    RESET_FAKE(set_middle_hex_digit);
+    RESET_FAKE(set_left_hex_digit);
+}
+
+void SevenSegDisplaysTest::TearDown()
+{
+    ASSERT_TRUE(left_seven_seg_display != NULL);
+    ASSERT_TRUE(middle_seven_seg_display != NULL);
+    ASSERT_TRUE(right_seven_seg_display != NULL);
+    ASSERT_TRUE(seven_seg_displays != NULL);
+
+    App_SevenSegDisplay_Destroy(left_seven_seg_display);
+    App_SevenSegDisplay_Destroy(middle_seven_seg_display);
+    App_SevenSegDisplay_Destroy(right_seven_seg_display);
+    App_SevenSegDisplays_Destroy(seven_seg_displays);
+
+    left_seven_seg_display   = NULL;
+    middle_seven_seg_display = NULL;
+    right_seven_seg_display  = NULL;
+    seven_seg_displays       = NULL;
+}
 
 TEST_F(SevenSegDisplaysTest, set_one_hexadecimal_digit)
 {
